LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/subresource-test/xhr-messaging-expected.txt

XHR messaging - cross-origin XHR request

If you can read JSON encoded HTTP request headers of the XHR below, the messaging works as expected.

http://127.0.0.1:8800/common/security-features/subresource/xhr.py:

{
  "origin": "http://localhost:8800",
  "accept-language": "fr-fr",
  "accept-encoding": "gzip, deflate",
  "host": "127.0.0.1:8800",
  "accept": "*/*",
  "user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_5) AppleWebKit/605.1.15 (KHTML, like Gecko)",
  "connection": "keep-alive",
  "referer": "http://localhost:8800/referrer-policy/generic/subresource-test/xhr-messaging.html"
}


PASS XHR is responding with HTTP headers
